[ADD] (image) 기본 테이블 및 repository 추가

+++ src/main/java/modak/camping/opendata/Base.java
@@ -14,7 +14,7 @@ import javax.persistence.*;
 @AllArgsConstructor
 public class Base {
     @Id
-    private int contentId;
+    private Long contentId;
     private String doNm;
     @Builder.Default
     private String mangeDivNm = "";
